requirements.txt.txt
scikit-learn==1.2.2
scipy==1.10.1
pandas==1.5.2
streamlit==1.19.0
matplotlib==3.4.3
numpy==1.21.5
yfinance==0.2.12
Pillow==9.3.0